docker/flask-postgres-compose-example/backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
Werkzeug==3.0.2
zipp==3.18.1
